[exporterhelper] Move NewLogsRequest to exporterhelper/internal
