+++ exporter/exporterhelper/internal/new_request.go
@@ -1,0 +1,69 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package internal // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal"
+
+import (
+	"context"
+
+	"go.opentelemetry.io/collector/consumer"
+	"go.opentelemetry.io/collector/consumer/consumererror"
+	"go.opentelemetry.io/collector/exporter"
+	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/request"
+	"go.opentelemetry.io/collector/pdata/plog"
+	"go.opentelemetry.io/collector/pipeline"
+	"go.uber.org/zap"
+)
+
+type logsExporter struct {
+	*BaseExporter
+	consumer.Logs
+}
+
+// NewLogsRequest creates new logs exporter based on custom LogsConverter and Sender.
+// Experimental: This API is at the early stage of development and may change without backward compatibility
+// until https://github.com/open-telemetry/opentelemetry-collector/issues/8122 is resolved.
+func NewLogsRequest(
+	_ context.Context,
+	set exporter.Settings,
+	converter request.RequestConverterFunc[plog.Logs],
+	pusher request.RequestConsumeFunc,
+	options ...Option,
+) (exporter.Logs, error) {
+	if set.Logger == nil {
+		return nil, errNilLogger
+	}
+
+	if converter == nil {
+		return nil, errNilLogsConverter
+	}
+
+	if pusher == nil {
+		return nil, errNilConsumeRequest
+	}
+
+	be, err := NewBaseExporter(set, pipeline.SignalLogs, pusher, options...)
+	if err != nil {
+		return nil, err
+	}
+
+	lc, err := consumer.NewLogs(newConsumeLogs(converter, be, set.Logger), be.ConsumerOptions...)
+	if err != nil {
+		return nil, err
+	}
+
+	return &logsExporter{BaseExporter: be, Logs: lc}, nil
+}
+
+func newConsumeLogs(converter request.RequestConverterFunc[plog.Logs], be *BaseExporter, logger *zap.Logger) consumer.ConsumeLogsFunc {
+	return func(ctx context.Context, ld plog.Logs) error {
+		req, err := converter(ctx, ld)
+		if err != nil {
+			logger.Error("Failed to convert logs. Dropping data.",
+				zap.Int("dropped_log_records", ld.LogRecordCount()),
+				zap.Error(err))
+			return consumererror.NewPermanent(err)
+		}
+		return be.Send(ctx, req)
+	}
+}
