+++ exporter/exporterhelper/logs_test.go
@@ -71,46 +71,15 @@ func TestLogs_NilLogger(t *testing.T) {
 	require.Equal(t, errNilLogger, err)
 }
 
-func TestLogsRequest_NilLogger(t *testing.T) {
-	le, err := NewLogsRequest(context.Background(), exporter.Settings{}, requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request]())
-	require.Nil(t, le)
-	require.Equal(t, errNilLogger, err)
-}
-
 func TestLogs_NilPushLogsData(t *testing.T) {
 	le, err := NewLogs(context.Background(), exportertest.NewNopSettings(exportertest.NopType), &fakeLogsConfig, nil)
 	require.Nil(t, le)
 	require.Equal(t, errNilPushLogs, err)
 }
 
-func TestLogsRequest_NilLogsConverter(t *testing.T) {
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType), nil, sendertest.NewNopSenderFunc[Request]())
-	require.Nil(t, le)
-	require.Equal(t, errNilLogsConverter, err)
-}
-
-func TestLogsRequest_NilPushLogsData(t *testing.T) {
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType), requesttest.RequestFromLogsFunc(nil), nil)
-	require.Nil(t, le)
-	require.Equal(t, errNilConsumeRequest, err)
-}
-
 func TestLogs_Default(t *testing.T) {
 	ld := plog.NewLogs()
 	le, err := NewLogs(context.Background(), exportertest.NewNopSettings(exportertest.NopType), &fakeLogsConfig, newPushLogsData(nil))
-	assert.NotNil(t, le)
-	require.NoError(t, err)
-
-	assert.Equal(t, consumer.Capabilities{MutatesData: false}, le.Capabilities())
-	assert.NoError(t, le.Start(context.Background(), componenttest.NewNopHost()))
-	assert.NoError(t, le.ConsumeLogs(context.Background(), ld))
-	assert.NoError(t, le.Shutdown(context.Background()))
-}
-
-func TestLogsRequest_Default(t *testing.T) {
-	ld := plog.NewLogs()
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request]())
 	assert.NotNil(t, le)
 	require.NoError(t, err)
 
@@ -129,40 +98,10 @@ func TestLogs_WithCapabilities(t *testing.T) {
 	assert.Equal(t, capabilities, le.Capabilities())
 }
 
-func TestLogsRequest_WithCapabilities(t *testing.T) {
-	capabilities := consumer.Capabilities{MutatesData: true}
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request](), WithCapabilities(capabilities))
-	require.NoError(t, err)
-	require.NotNil(t, le)
-
-	assert.Equal(t, capabilities, le.Capabilities())
-}
-
 func TestLogs_Default_ReturnError(t *testing.T) {
 	ld := plog.NewLogs()
 	want := errors.New("my_error")
 	le, err := NewLogs(context.Background(), exportertest.NewNopSettings(exportertest.NopType), &fakeLogsConfig, newPushLogsData(want))
-	require.NoError(t, err)
-	require.NotNil(t, le)
-	require.Equal(t, want, le.ConsumeLogs(context.Background(), ld))
-}
-
-func TestLogsRequest_Default_ConvertError(t *testing.T) {
-	ld := plog.NewLogs()
-	want := errors.New("convert_error")
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromLogsFunc(want), sendertest.NewNopSenderFunc[Request]())
-	require.NoError(t, err)
-	require.NotNil(t, le)
-	require.Equal(t, consumererror.NewPermanent(want), le.ConsumeLogs(context.Background(), ld))
-}
-
-func TestLogsRequest_Default_ExportError(t *testing.T) {
-	ld := plog.NewLogs()
-	want := errors.New("export_error")
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromLogsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
 	require.NoError(t, err)
 	require.NotNil(t, le)
 	require.Equal(t, want, le.ConsumeLogs(context.Background(), ld))
@@ -279,7 +218,7 @@ func TestLogsRequest_WithRecordMetrics(t *testing.T) {
 	tt := componenttest.NewTelemetry()
 	t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
 
-	le, err := NewLogsRequest(context.Background(),
+	le, err := internal.NewLogsRequest(context.Background(),
 		exporter.Settings{ID: fakeLogsName, TelemetrySettings: tt.NewTelemetrySettings(), BuildInfo: component.NewDefaultBuildInfo()},
 		requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request]())
 	require.NoError(t, err)
@@ -305,7 +244,7 @@ func TestLogsRequest_WithRecordMetrics_ExportError(t *testing.T) {
 	tt := componenttest.NewTelemetry()
 	t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
 
-	le, err := NewLogsRequest(context.Background(), exporter.Settings{ID: fakeLogsName, TelemetrySettings: tt.NewTelemetrySettings(), BuildInfo: component.NewDefaultBuildInfo()},
+	le, err := internal.NewLogsRequest(context.Background(), exporter.Settings{ID: fakeLogsName, TelemetrySettings: tt.NewTelemetrySettings(), BuildInfo: component.NewDefaultBuildInfo()},
 		requesttest.RequestFromLogsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
 	require.NoError(t, err)
 	require.NotNil(t, le)
@@ -333,7 +272,7 @@ func TestLogsRequest_WithSpan(t *testing.T) {
 	otel.SetTracerProvider(set.TracerProvider)
 	defer otel.SetTracerProvider(nooptrace.NewTracerProvider())
 
-	le, err := NewLogsRequest(context.Background(), set, requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request]())
+	le, err := internal.NewLogsRequest(context.Background(), set, requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request]())
 	require.NoError(t, err)
 	require.NotNil(t, le)
 	checkWrapSpanForLogs(t, sr, set.TracerProvider.Tracer("test"), le, nil)
@@ -361,7 +300,7 @@ func TestLogsRequest_WithSpan_ReturnError(t *testing.T) {
 	defer otel.SetTracerProvider(nooptrace.NewTracerProvider())
 
 	want := errors.New("my_error")
-	le, err := NewLogsRequest(context.Background(), set, requesttest.RequestFromLogsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
+	le, err := internal.NewLogsRequest(context.Background(), set, requesttest.RequestFromLogsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
 	require.NoError(t, err)
 	require.NotNil(t, le)
 	checkWrapSpanForLogs(t, sr, set.TracerProvider.Tracer("test"), le, want)
@@ -379,36 +318,11 @@ func TestLogs_WithShutdown(t *testing.T) {
 	assert.True(t, shutdownCalled)
 }
 
-func TestLogsRequest_WithShutdown(t *testing.T) {
-	shutdownCalled := false
-	shutdown := func(context.Context) error { shutdownCalled = true; return nil }
-
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request](), WithShutdown(shutdown))
-	assert.NotNil(t, le)
-	assert.NoError(t, err)
-
-	assert.NoError(t, le.Shutdown(context.Background()))
-	assert.True(t, shutdownCalled)
-}
-
 func TestLogs_WithShutdown_ReturnError(t *testing.T) {
 	want := errors.New("my_error")
 	shutdownErr := func(context.Context) error { return want }
 
 	le, err := NewLogs(context.Background(), exportertest.NewNopSettings(exportertest.NopType), &fakeLogsConfig, newPushLogsData(nil), WithShutdown(shutdownErr))
-	assert.NotNil(t, le)
-	require.NoError(t, err)
-
-	assert.Equal(t, want, le.Shutdown(context.Background()))
-}
-
-func TestLogsRequest_WithShutdown_ReturnError(t *testing.T) {
-	want := errors.New("my_error")
-	shutdownErr := func(context.Context) error { return want }
-
-	le, err := NewLogsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromLogsFunc(nil), sendertest.NewNopSenderFunc[Request](), WithShutdown(shutdownErr))
 	assert.NotNil(t, le)
 	require.NoError(t, err)
