+++ exporter/exporterhelper/request.go
@@ -4,10 +4,7 @@
 package exporterhelper // import "go.opentelemetry.io/collector/exporter/exporterhelper"
 
 import (
-	"context"
-
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/request"
-	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/sender"
 )
 
 // Request represents a single request that can be sent to an external endpoint.
@@ -26,11 +23,11 @@ type RequestErrorHandler = request.ErrorHandler
 // RequestConverterFunc converts pdata telemetry into a user-defined Request.
 // Experimental: This API is at the early stage of development and may change without backward compatibility
 // until https://github.com/open-telemetry/opentelemetry-collector/issues/8122 is resolved.
-type RequestConverterFunc[T any] func(context.Context, T) (Request, error)
+type RequestConverterFunc[T any] = request.RequestConverterFunc[T]
 
 // RequestConsumeFunc processes the request. After the function returns, the request is no longer accessible,
 // and accessing it is considered undefined behavior.
-type RequestConsumeFunc = sender.SendFunc[Request]
+type RequestConsumeFunc = request.RequestConsumeFunc
 
 // RequestSizer is an interface that returns the size of the given request.
 type RequestSizer = request.Sizer[Request]
